Move some OAuth constants to data layer, according to last CR suggestion
